requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.0.14
djangorestframework==3.16.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mssql-django==1.5
pymssql==2.3.4
pyodbc==5.2.0
pytz==2025.2
requests==2.32.3
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
